add: subject

+++ server/src/controllers/cluster/removeById.js
@@ -7,6 +7,7 @@ const { ctrlWrapper } = require('../../decorators');
 const { restrictedAccess } = require('../../utils');
 
 const removeById = ctrlWrapper(async (req, res) => {
+  const owner = req.user._id;
   let { id } = req.params;
   id = id.includes('&') ? id.split('&') : [id];
 
@@ -30,9 +31,9 @@ const removeById = ctrlWrapper(async (req, res) => {
   if (!clusterCount) throw HttpError(403, 'Failed to delete cluster');
   // groups
   for (let i = 0; i < groups.length; i += 1) {
-    const clusters = await Cluster.find({ group: groups[i] });
+    const clusters = await Cluster.find({ owner, group: groups[i] });
     if (clusters.length === 0) {
-      const delGroup = await ClusterGroup.findOneAndDelete({ clusterGroup: groups[i] });
+      const delGroup = await ClusterGroup.findOneAndDelete({ owner, clusterGroup: groups[i] });
       if (!delGroup) throw HttpError(403, 'Failed to delete group');
       groupCount += 1;
     }
